Move serialized lambda extraction into ImplementationSnapshot
