+++ subprojects/core/src/main/java/org/gradle/api/internal/AbstractTask.java
@@ -80,7 +80,6 @@ import org.gradle.internal.resources.SharedResource;
 import org.gradle.internal.scripts.ScriptOrigin;
 import org.gradle.internal.service.ServiceRegistry;
 import org.gradle.internal.snapshot.impl.ImplementationSnapshot;
-import org.gradle.internal.snapshot.impl.LambdaImplementationSnapshot;
 import org.gradle.util.Path;
 import org.gradle.util.internal.ConfigureUtil;
 import org.gradle.work.DisableCachingByDefault;
@@ -89,10 +88,6 @@ import javax.annotation.Nullable;
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 import java.io.File;
-import java.io.Serializable;
-import java.lang.invoke.SerializedLambda;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -807,36 +802,8 @@ public abstract class AbstractTask implements TaskInternal, DynamicObjectAware {
         @Override
         public ImplementationSnapshot getActionImplementation(ClassLoaderHierarchyHasher hasher) {
             HashCode classLoaderHash = hasher.getClassLoaderHash(action.getClass().getClassLoader());
-            ImplementationSnapshot snapshot = ImplementationSnapshot.of(AbstractTask.getActionClassName(action), classLoaderHash);
-            if (snapshot instanceof LambdaImplementationSnapshot) {
-                Optional<SerializedLambda> serializedLambda = serializedLambdaFor(action);
-                return serializedLambda.map(it -> ImplementationSnapshot.of(it, classLoaderHash))
-                    .orElse(snapshot);
-            }
-            return snapshot;
-        }
-
-        private static Optional<SerializedLambda> serializedLambdaFor(Object lambda) {
-            if (!(lambda instanceof Serializable)) {
-                return Optional.empty();
-            }
-            for (Class<?> lambdaClass = lambda.getClass(); lambdaClass != null; lambdaClass = lambdaClass.getSuperclass()) {
-                try {
-                    Method replaceMethod = lambdaClass.getDeclaredMethod("writeReplace");
-                    replaceMethod.setAccessible(true);
-                    Object serializedForm = replaceMethod.invoke(lambda);
-                    if (serializedForm instanceof SerializedLambda) {
-                        return Optional.of((SerializedLambda) serializedForm);
-                    } else {
-                        return Optional.empty();
-                    }
-                } catch (NoSuchMethodException e) {
-                    // continue
-                } catch (InvocationTargetException | IllegalAccessException e) {
-                    return Optional.empty();
-                }
-            }
-            return Optional.empty();
+            String actionClassName = AbstractTask.getActionClassName(action);
+            return ImplementationSnapshot.of(actionClassName, action, classLoaderHash);
         }
 
         @Override
